more git ignore stuff in integration test
