--- test/integration/excludeMenu/expected/repo/.git_keep/COMMIT_EDITMSG
@@ -1,1 +1,0 @@
-Initial commit
