--- test/integration/excludeMenu/expected/repo/.git_keep/HEAD
@@ -1,1 +1,0 @@
-ref: refs/heads/master
