--- test/integration/excludeMenu/expected/repo/.git_keep/description
@@ -1,1 +1,0 @@
-Unnamed repository; edit this file 'description' to name the repository.
