--- test/integration/excludeMenu/expected/repo/.git_keep/info/exclude
@@ -1,9 +1,0 @@
-# git ls-files --others --exclude-from=.git/info/exclude
-# Lines that start with '#' are comments.
-# For a project mostly in C, the following would be a good set of
-# exclude patterns (uncomment them if you want to use them):
-# *.[oa]
-# *~
-.DS_Store
-
-myfile1
