--- test/integration/excludeMenu/expected/repo/.git_keep/logs/refs/heads/master
@@ -1,1 +1,0 @@
-0000000000000000000000000000000000000000 129cdae0c4ccd050e8398bcb18b2ce1e4a5626f9 CI <CI@example.com> 1657012793 +1000	commit (initial): Initial commit
