--- test/integration/excludeMenu/expected/repo/.git_keep/refs/heads/master
@@ -1,1 +1,0 @@
-129cdae0c4ccd050e8398bcb18b2ce1e4a5626f9
