--- test/integration/excludeMenu/expected/repo/myfile1
@@ -1,1 +1,0 @@
-test1
