--- test/integration/excludeMenu/test.json
@@ -1,4 +1,0 @@
-{
-    "description": "In this test a file is added to .git/info/exclude using the ignore or exclude menu",
-    "speed": 5
-}
